--- confmap/provider/eeciprovider/provider_test.go
@@ -1,20 +1,0 @@
-// Copyright The OpenTelemetry Authors
-// SPDX-License-Identifier: Apache-2.0
-
-package eecprovider
-
-import (
-	"context"
-	"testing"
-
-	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
-
-	"go.opentelemetry.io/collector/confmap/confmaptest"
-)
-
-func TestSupportedScheme(t *testing.T) {
-	fp := NewFactory().Create(confmaptest.NewNopProviderSettings())
-	assert.Equal(t, "eeci", fp.Scheme())
-	require.NoError(t, fp.Shutdown(context.Background()))
-}
