+++ confmap/provider/eecprovider/README.md
@@ -13,7 +13,7 @@ Collector to be configured with the Dynatrace Extensions Execution Controller.
 This confmap Provider can be configured by pasing query parameter-formatted values
 inside the fragment of the URL given to the config flag. For example:
 
-```
+```text
 dynatrace-otel-collector --config=eec://my.eec.host:31098#refresh-interval=5s&auth-header=Api-Token&auth-file=/var/private/token.key
 ```
 
@@ -25,3 +25,4 @@ dynatrace-otel-collector --config=eec://my.eec.host:31098#refresh-interval=5s&au
 | auth-file | None | A filepath containing a plaintext version of an API token or other key to be used in an HTTP header to authenticate with the EEC. |
 | auth-header | Api-Token | A header that will be used to authenticate the provider with the EEC. |
 | refresh-interval | 10s | A time duration that defines how frequently the provider should check the given URL for updates. |
+| insecure | false | If set to "true", use HTTP for the connection to the server. If unset or set to "false", use HTTPS. |
