+++ confmap/provider/eecprovider/watcher.go
