+++ confmap/provider/eecprovider/watcher_test.go
