--- confmap/provider/go.mod
@@ -1,23 +1,0 @@
-module github.com/Dynatrace/dynatrace-otel-collector/confmap/provider
-
-go 1.21
-
-require (
-	github.com/stretchr/testify v1.9.0
-	go.opentelemetry.io/collector/confmap v0.99.0
-	go.uber.org/goleak v1.3.0
-	gopkg.in/yaml.v3 v3.0.1
-)
-
-require (
-	github.com/davecgh/go-spew v1.1.1 // indirect
-	github.com/go-viper/mapstructure/v2 v2.0.0-alpha.1 // indirect
-	github.com/knadh/koanf/maps v0.1.1 // indirect
-	github.com/knadh/koanf/providers/confmap v0.1.0 // indirect
-	github.com/knadh/koanf/v2 v2.1.1 // indirect
-	github.com/mitchellh/copystructure v1.2.0 // indirect
-	github.com/mitchellh/reflectwalk v1.0.2 // indirect
-	github.com/pmezard/go-difflib v1.0.0 // indirect
-	go.uber.org/multierr v1.11.0 // indirect
-	go.uber.org/zap v1.27.0 // indirect
-)
