--- confmap/provider/internal/configurablehttpprovider/provider_test.go
@@ -1,503 +1,0 @@
-// Copyright The OpenTelemetry Authors
-// SPDX-License-Identifier: Apache-2.0
-
-package configurablehttpprovider
-
-import (
-	"context"
-	"crypto/rand"
-	"crypto/rsa"
-	"crypto/tls"
-	"crypto/x509"
-	"crypto/x509/pkix"
-	"encoding/pem"
-	"fmt"
-	"math/big"
-	"net/http"
-	"net/http/httptest"
-	"net/url"
-	"os"
-	"path"
-	"strings"
-	"sync"
-	"sync/atomic"
-	"testing"
-	"time"
-
-	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
-
-	"go.opentelemetry.io/collector/confmap"
-	"go.opentelemetry.io/collector/confmap/confmaptest"
-)
-
-func newConfigurableHTTPProvider(scheme SchemeType, set confmap.ProviderSettings) *provider {
-	return New(scheme, set).(*provider)
-}
-
-func answerGet(w http.ResponseWriter, _ *http.Request) {
-	f, err := os.ReadFile("./testdata/otel-config.yaml")
-	if err != nil {
-		w.WriteHeader(http.StatusNotFound)
-		_, innerErr := w.Write([]byte("Cannot find the config file"))
-		if innerErr != nil {
-			fmt.Println("Write failed: ", innerErr)
-		}
-		return
-	}
-	w.WriteHeader(http.StatusOK)
-	_, err = w.Write(f)
-	if err != nil {
-		fmt.Println("Write failed: ", err)
-	}
-}
-
-// Generate a self signed certificate specific for the tests. Based on
-// https://go.dev/src/crypto/tls/generate_cert.go
-func generateCertificate(hostname string) (cert string, key string, err error) {
-	priv, err := rsa.GenerateKey(rand.Reader, 2048)
-
-	if err != nil {
-		return "", "", fmt.Errorf("Failed to generate private key: %w", err)
-	}
-
-	keyUsage := x509.KeyUsageDigitalSignature | x509.KeyUsageKeyEncipherment | x509.KeyUsageCertSign
-	notBefore := time.Now()
-	notAfter := notBefore.Add(time.Hour * 12)
-	serialNumberLimit := new(big.Int).Lsh(big.NewInt(1), 128)
-	serialNumber, err := rand.Int(rand.Reader, serialNumberLimit)
-
-	if err != nil {
-		return "", "", fmt.Errorf("Failed to generate serial number: %w", err)
-	}
-
-	template := x509.Certificate{
-		SerialNumber: serialNumber,
-		Subject: pkix.Name{
-			Organization: []string{"Httpprovider Co"},
-		},
-		NotBefore:             notBefore,
-		NotAfter:              notAfter,
-		KeyUsage:              keyUsage,
-		ExtKeyUsage:           []x509.ExtKeyUsage{x509.ExtKeyUsageServerAuth},
-		BasicConstraintsValid: true,
-		IsCA:                  true,
-		DNSNames:              []string{hostname},
-	}
-	derBytes, err := x509.CreateCertificate(rand.Reader, &template, &template, &priv.PublicKey, priv)
-
-	if err != nil {
-		return "", "", fmt.Errorf("Failed to create certificate: %w", err)
-	}
-
-	certOut, err := os.CreateTemp("", "cert*.pem")
-	if err != nil {
-		return "", "", fmt.Errorf("Failed to open cert.pem for writing: %w", err)
-	}
-
-	if err = pem.Encode(certOut, &pem.Block{Type: "CERTIFICATE", Bytes: derBytes}); err != nil {
-		return "", "", fmt.Errorf("Failed to write data to cert.pem: %w", err)
-	}
-
-	if err = certOut.Close(); err != nil {
-		return "", "", fmt.Errorf("Error closing cert.pem: %w", err)
-	}
-
-	keyOut, err := os.CreateTemp("", "key*.pem")
-
-	if err != nil {
-		return "", "", fmt.Errorf("Failed to open key.pem for writing: %w", err)
-	}
-
-	privBytes, err := x509.MarshalPKCS8PrivateKey(priv)
-
-	if err != nil {
-		return "", "", fmt.Errorf("Unable to marshal private key: %w", err)
-	}
-
-	if err := pem.Encode(keyOut, &pem.Block{Type: "PRIVATE KEY", Bytes: privBytes}); err != nil {
-		return "", "", fmt.Errorf("Failed to write data to key.pem: %w", err)
-	}
-
-	if err := keyOut.Close(); err != nil {
-		return "", "", fmt.Errorf("Error closing key.pem: %w", err)
-	}
-
-	return certOut.Name(), keyOut.Name(), nil
-}
-
-func TestFunctionalityDownloadFileHTTP(t *testing.T) {
-	fp := newConfigurableHTTPProvider(EECIScheme, confmaptest.NewNopProviderSettings())
-	ts := httptest.NewServer(http.HandlerFunc(answerGet))
-	defer ts.Close()
-	_, err := fp.Retrieve(context.Background(), strings.Replace(ts.URL, "http", "eeci", 1), nil)
-	assert.NoError(t, err)
-	assert.NoError(t, fp.Shutdown(context.Background()))
-}
-
-func TestFunctionalityDownloadFileHTTPS(t *testing.T) {
-	certPath, keyPath, err := generateCertificate("localhost")
-	assert.NoError(t, err)
-
-	invalidCert, err := os.CreateTemp("", "cert*.crt")
-	assert.NoError(t, err)
-	_, err = invalidCert.Write([]byte{0, 1, 2})
-	assert.NoError(t, err)
-
-	cert, err := tls.LoadX509KeyPair(certPath, keyPath)
-	assert.NoError(t, err)
-	ts := httptest.NewUnstartedServer(http.HandlerFunc(answerGet))
-	ts.TLS = &tls.Config{Certificates: []tls.Certificate{cert}}
-	ts.StartTLS()
-
-	defer os.Remove(certPath)
-	defer os.Remove(keyPath)
-	defer os.Remove(invalidCert.Name())
-	defer ts.Close()
-
-	tests := []struct {
-		testName               string
-		certPath               string
-		hostName               string
-		useCertificate         bool
-		skipHostnameValidation bool
-		shouldError            bool
-	}{
-		{
-			testName:               "Test valid certificate and name",
-			certPath:               certPath,
-			hostName:               "localhost",
-			useCertificate:         true,
-			skipHostnameValidation: false,
-			shouldError:            false,
-		},
-		{
-			testName:               "Test valid certificate with invalid name",
-			certPath:               certPath,
-			hostName:               "127.0.0.1",
-			useCertificate:         true,
-			skipHostnameValidation: false,
-			shouldError:            true,
-		},
-		{
-			testName:               "Test valid certificate with invalid name, skip validation",
-			certPath:               certPath,
-			hostName:               "127.0.0.1",
-			useCertificate:         true,
-			skipHostnameValidation: true,
-			shouldError:            false,
-		},
-		{
-			testName:               "Test no certificate should fail",
-			certPath:               certPath,
-			hostName:               "localhost",
-			useCertificate:         false,
-			skipHostnameValidation: false,
-			shouldError:            true,
-		},
-		{
-			testName:               "Test invalid cert",
-			certPath:               invalidCert.Name(),
-			hostName:               "localhost",
-			useCertificate:         true,
-			skipHostnameValidation: false,
-			shouldError:            true,
-		},
-		{
-			testName:               "Test no cert",
-			certPath:               "no_certificate",
-			hostName:               "localhost",
-			useCertificate:         true,
-			skipHostnameValidation: false,
-			shouldError:            true,
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.testName, func(t *testing.T) {
-			fp := newConfigurableHTTPProvider(EECScheme, confmaptest.NewNopProviderSettings())
-			// Parse url of the test server to get the port number.
-			tsURL, err := url.Parse(ts.URL)
-			require.NoError(t, err)
-			if tt.useCertificate {
-				fp.caCertPath = tt.certPath
-			}
-			fp.insecureSkipVerify = tt.skipHostnameValidation
-			_, err = fp.Retrieve(context.Background(), fmt.Sprintf("eec://%s:%s", tt.hostName, tsURL.Port()), nil)
-			if tt.shouldError {
-				assert.Error(t, err)
-			} else {
-				assert.NoError(t, err)
-			}
-		})
-	}
-}
-
-func TestUnsupportedScheme(t *testing.T) {
-	fp := New(EECIScheme, confmaptest.NewNopProviderSettings())
-	_, err := fp.Retrieve(context.Background(), "eec://...", nil)
-	assert.Error(t, err)
-	assert.NoError(t, fp.Shutdown(context.Background()))
-
-	fp = New(EECScheme, confmaptest.NewNopProviderSettings())
-	_, err = fp.Retrieve(context.Background(), "eeci://...", nil)
-	assert.Error(t, err)
-	assert.NoError(t, fp.Shutdown(context.Background()))
-}
-
-func TestEmptyURI(t *testing.T) {
-	fp := New(EECIScheme, confmaptest.NewNopProviderSettings())
-	ts := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, _ *http.Request) {
-		w.WriteHeader(http.StatusBadRequest)
-	}))
-	defer ts.Close()
-	_, err := fp.Retrieve(context.Background(), strings.Replace(ts.URL, "http", "eeci", 1), nil)
-	require.Error(t, err)
-	require.NoError(t, fp.Shutdown(context.Background()))
-}
-
-func TestRetrieveFromShutdownServer(t *testing.T) {
-	fp := New(EECIScheme, confmaptest.NewNopProviderSettings())
-	ts := httptest.NewServer(http.HandlerFunc(func(http.ResponseWriter, *http.Request) {}))
-	ts.Close()
-	_, err := fp.Retrieve(context.Background(), strings.Replace(ts.URL, "http", "eeci", 1), nil)
-	assert.Error(t, err)
-	require.NoError(t, fp.Shutdown(context.Background()))
-}
-
-func TestNonExistent(t *testing.T) {
-	fp := New(EECIScheme, confmaptest.NewNopProviderSettings())
-	ts := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, _ *http.Request) {
-		w.WriteHeader(http.StatusNotFound)
-	}))
-	defer ts.Close()
-	_, err := fp.Retrieve(context.Background(), strings.Replace(ts.URL, "http", "eeci", 1), nil)
-	assert.Error(t, err)
-	require.NoError(t, fp.Shutdown(context.Background()))
-}
-
-func TestInvalidYAML(t *testing.T) {
-	fp := New(EECIScheme, confmaptest.NewNopProviderSettings())
-	ts := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, _ *http.Request) {
-		w.WriteHeader(http.StatusOK)
-		_, err := w.Write([]byte("wrong : ["))
-		if err != nil {
-			fmt.Println("Write failed: ", err)
-		}
-	}))
-	defer ts.Close()
-	_, err := fp.Retrieve(context.Background(), strings.Replace(ts.URL, "http", "eeci", 1), nil)
-	assert.Error(t, err)
-	require.NoError(t, fp.Shutdown(context.Background()))
-}
-
-func TestScheme(t *testing.T) {
-	fp := New(EECIScheme, confmaptest.NewNopProviderSettings())
-	assert.Equal(t, "eeci", fp.Scheme())
-	require.NoError(t, fp.Shutdown(context.Background()))
-}
-
-func TestValidateProviderScheme(t *testing.T) {
-	assert.NoError(t, confmaptest.ValidateProviderScheme(New(EECIScheme, confmaptest.NewNopProviderSettings())))
-}
-
-func TestInvalidTransport(t *testing.T) {
-	fp := New("foo", confmaptest.NewNopProviderSettings())
-
-	_, err := fp.Retrieve(context.Background(), "foo://..", nil)
-	assert.Error(t, err)
-}
-
-func TestNoReloadIfConfigIsSame(t *testing.T) {
-	count := &atomic.Uint32{}
-	answerWithCount := func(w http.ResponseWriter, _ *http.Request) {
-		f, err := os.ReadFile("./testdata/otel-config.yaml")
-		if err != nil {
-			w.WriteHeader(http.StatusNotFound)
-			_, innerErr := w.Write([]byte("Cannot find the config file"))
-			if innerErr != nil {
-				fmt.Println("Write failed: ", innerErr)
-			}
-			return
-		}
-		w.WriteHeader(http.StatusOK)
-		_, err = w.Write(f)
-		if err != nil {
-			fmt.Println("Write failed: ", err)
-		}
-		count.Add(1)
-	}
-	called := &atomic.Bool{}
-	watcherFunc := func(_ *confmap.ChangeEvent) {
-		called.Store(true)
-	}
-	fp := newConfigurableHTTPProvider(EECIScheme, confmaptest.NewNopProviderSettings())
-	ts := httptest.NewServer(http.HandlerFunc(answerWithCount))
-	defer ts.Close()
-	uri, err := url.Parse(ts.URL)
-	require.NoError(t, err)
-	params := url.Values{}
-	params.Set(RefreshInterval, "10ms")
-	uri.Fragment = params.Encode()
-	uri.Scheme = "eeci"
-
-	_, err = fp.Retrieve(context.Background(), uri.String(), watcherFunc)
-	require.NoError(t, err)
-	require.Eventually(t, func() bool {
-		return count.Load() > 2
-	}, time.Second*3, time.Millisecond*50)
-	require.False(t, called.Load())
-	assert.NoError(t, fp.Shutdown(context.Background()))
-}
-
-func TestReloadIfConfigChanges(t *testing.T) {
-	count := &atomic.Uint32{}
-	answerWithCount := func(w http.ResponseWriter, _ *http.Request) {
-		configFile := "./testdata/otel-config.yaml"
-		if count.Load() > 2 {
-			configFile = "./testdata/otel-config-updated.yaml"
-		}
-		f, err := os.ReadFile(configFile)
-		if err != nil {
-			w.WriteHeader(http.StatusNotFound)
-			_, innerErr := w.Write([]byte("Cannot find the config file"))
-			if innerErr != nil {
-				fmt.Println("Write failed: ", innerErr)
-			}
-			return
-		}
-		w.WriteHeader(http.StatusOK)
-		_, err = w.Write(f)
-		if err != nil {
-			fmt.Println("Write failed: ", err)
-		}
-		count.Add(1)
-	}
-	called := &atomic.Bool{}
-	watcherFunc := func(_ *confmap.ChangeEvent) {
-		called.Store(true)
-	}
-	fp := newConfigurableHTTPProvider(EECIScheme, confmaptest.NewNopProviderSettings())
-	ts := httptest.NewServer(http.HandlerFunc(answerWithCount))
-	defer ts.Close()
-	uri, err := url.Parse(ts.URL)
-	require.NoError(t, err)
-	params := url.Values{}
-	params.Set(RefreshInterval, "10ms")
-	uri.Fragment = params.Encode()
-	uri.Scheme = "eeci"
-
-	_, err = fp.Retrieve(context.Background(), uri.String(), watcherFunc)
-	require.NoError(t, err)
-	require.Eventually(t, func() bool {
-		return count.Load() > 2
-	}, time.Second*3, time.Millisecond*50)
-	require.True(t, called.Load())
-	assert.NoError(t, fp.Shutdown(context.Background()))
-}
-
-func TestContinuesRetryingOnRefreshError(t *testing.T) {
-	count := &atomic.Uint32{}
-	answerWithCount := func(w http.ResponseWriter, _ *http.Request) {
-		if count.Load() > 2 {
-			w.WriteHeader(http.StatusInternalServerError)
-			_, err := w.Write(nil)
-			if err != nil {
-				fmt.Println("Write failed: ", err)
-			}
-			count.Add(1)
-			return
-		}
-
-		f, err := os.ReadFile("./testdata/otel-config.yaml")
-		if err != nil {
-			w.WriteHeader(http.StatusNotFound)
-			_, innerErr := w.Write([]byte("Cannot find the config file"))
-			if innerErr != nil {
-				fmt.Println("Write failed: ", innerErr)
-			}
-			return
-		}
-		w.WriteHeader(http.StatusOK)
-		_, err = w.Write(f)
-		if err != nil {
-			fmt.Println("Write failed: ", err)
-		}
-		count.Add(1)
-	}
-	called := &atomic.Bool{}
-	watcherFunc := func(_ *confmap.ChangeEvent) {
-		called.Store(true)
-	}
-	fp := newConfigurableHTTPProvider(EECIScheme, confmaptest.NewNopProviderSettings())
-	ts := httptest.NewServer(http.HandlerFunc(answerWithCount))
-	defer ts.Close()
-	uri, err := url.Parse(ts.URL)
-	require.NoError(t, err)
-	params := url.Values{}
-	params.Set(RefreshInterval, "10ms")
-	uri.Fragment = params.Encode()
-	uri.Scheme = "eeci"
-
-	_, err = fp.Retrieve(context.Background(), uri.String(), watcherFunc)
-	require.NoError(t, err)
-	require.Eventually(t, func() bool {
-		return count.Load() > 5
-	}, time.Second*3, time.Millisecond*20)
-	require.False(t, called.Load())
-	assert.NoError(t, fp.Shutdown(context.Background()))
-}
-
-func TestFragmentConfiguration(t *testing.T) {
-	wg := &sync.WaitGroup{}
-	authHeader := "Bearer-Token"
-	refreshInterval := "10h"
-
-	token := "mytoken"
-	tmpDir := t.TempDir()
-	file, err := os.Create(path.Join(tmpDir, "token.key"))
-	require.NoError(t, err)
-	file.Write([]byte(token))
-
-	answerWithConfig := func(w http.ResponseWriter, req *http.Request) {
-		f, err := os.ReadFile("./testdata/otel-config.yaml")
-		if err != nil {
-			w.WriteHeader(http.StatusNotFound)
-			_, innerErr := w.Write([]byte("Cannot find the config file"))
-			if innerErr != nil {
-				fmt.Println("Write failed: ", innerErr)
-			}
-			return
-		}
-		w.WriteHeader(http.StatusOK)
-		_, err = w.Write(f)
-		if err != nil {
-			fmt.Println("Write failed: ", err)
-		}
-
-		require.Equal(t, token, req.Header.Get(authHeader))
-
-		wg.Done()
-	}
-	called := &atomic.Bool{}
-	watcherFunc := func(_ *confmap.ChangeEvent) {
-		called.Store(true)
-	}
-	fp := newConfigurableHTTPProvider(EECIScheme, confmaptest.NewNopProviderSettings())
-	ts := httptest.NewServer(http.HandlerFunc(answerWithConfig))
-	defer ts.Close()
-	uri, err := url.Parse(ts.URL)
-	require.NoError(t, err)
-	params := url.Values{}
-	params.Set(AuthHeader, authHeader)
-	params.Set(AuthFile, file.Name())
-	params.Set(RefreshInterval, refreshInterval)
-	uri.Fragment = params.Encode()
-	uri.Scheme = "eeci"
-
-	wg.Add(1)
-	_, err = fp.Retrieve(context.Background(), uri.String(), watcherFunc)
-	require.NoError(t, err)
-	wg.Wait()
-	assert.NoError(t, fp.Shutdown(context.Background()))
-}
